Generación de Minutas/requirements.txt
openai>=1.52.0
streamlit>=1.36.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
pandas>=2.2.2
jinja2>=3.1.4
webvtt-py>=0.5.1
srt>=3.5.3
python-docx>=1.1.2
pypandoc>=1.13
rapidfuzz>=3.9.6
